% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_farms}
\alias{canada_farms}
\title{Canada Maple Farm Addresses}
\format{
\subsection{\code{canada_farms}}{

A dataframe with 15 rows and 4 columns:
\describe{
\item{farm}{Farm name}
\item{address}{Farm address}
\item{state}{Province}
\item{region}{Province region}
}
}
}
\usage{
canada_farms
}
\description{
Maple syrup farm addresses scraped from
\href{https://www.maplesyrupfarms.org/}{maplesyrupfarms.org}.
}
\keyword{datasets}
