% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnd.R
\name{download_ghcnd}
\alias{download_ghcnd}
\title{Download GHCNd files with given ids}
\usage{
download_ghcnd(ids, out_dir = paste0(getwd()))
}
\arguments{
\item{ids}{character vector of GHCNd station ids}

\item{out_dir}{name of the directory where the files will be downloaded to}
}
\description{
Download GHCNd files with given ids
}
\examples{
\dontrun{
# Download files for all U.S. stations
my_ids <- get_ghcnd_country_ids("US")
download_ghcnd(my_ids)
}
}
