% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farms_coords}
\alias{farms_coords}
\title{North America Maple Farm Addresses and Coordinates}
\format{
\subsection{\code{farms_coords}}{

A dataframe with 444 rows and 6 columns:
\describe{
\item{farm}{Farm name}
\item{address}{Farm address}
\item{state}{U.S. state/Canada province}
\item{region}{U.S. state/Canada province region}
\item{lat}{Latitude}
\item{long}{Longitude}
}
}
}
\usage{
farms_coords
}
\description{
Maple syrup farm addresses and coordinates scraped from
\href{https://www.maplesyrupfarms.org/}{maplesyrupfarms.org} and
\href{https://www.bonjourquebec.com/en-us/to-see-and-do/delicious-discoveries/sugar-shacks}{bonjourquebec.com}.
Since scraping the farm names was complicated, a few farm names are cut
off at the first hyphen. Farm addresses were geocoded with
\code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}}.
}
\keyword{datasets}
