% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnd.R
\name{get_ghcnd_country_ids}
\alias{get_ghcnd_country_ids}
\title{Get all GHCNd station ids for a given country}
\usage{
get_ghcnd_country_ids(country_code)
}
\arguments{
\item{country_code}{character vector of WMO country codes}
}
\value{
character vector of station ids
}
\description{
Get all GHCNd station ids for a given country
}
\examples{
# Get all U.S. country ids
get_ghcnd_country_ids("US")
}
