% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loca2_model_names.R
\name{get_loca2_model_names}
\alias{get_loca2_model_names}
\title{Get all LOCA2 model names}
\usage{
get_loca2_model_names()
}
\value{
character vector of length 27 with all valid model names
}
\description{
Get all LOCA2 model names
}
