% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghcnd_stations}
\alias{ghcnd_stations}
\title{GHCNd Station Data}
\format{
\subsection{\code{ghcnd_stations}}{

A dataframe with 129658 rows and 9 columns:
\describe{
\item{ID}{Station identification code}
\item{LATITUDE}{Station latitude in degrees}
\item{LONGITUDE}{Station longitude in degrees}
\item{ELEVATION}{Station elevation in meters}
\item{STATE}{U.S. postal code for the state (for U.S. stations only)}
\item{NAME}{Station name}
\item{GSN_FLAG}{Flag that indicates whether the station is part of
the GCOS Surface Network (GSN). Unique values are blank and "GSN"}
\item{HCN_CRN_FLAG}{Flag that indicates whether the station is part of the
U.S. Historical Climatology Network (HCN) or U.S. Climate Reference
Network (CRN). Unique values are blank, "HCN", and "CRN"}
\item{WMO_ID}{Station World Meteorological Organization (WMO) number}
}
}
}
\usage{
ghcnd_stations
}
\description{
GHCNd station data pulled from
\href{https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt}{ghcnd-stations.txt}.
}
\keyword{datasets}
