% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loca2_model_names}
\alias{loca2_model_names}
\title{LOCA2 Model Names}
\format{
\subsection{\code{loca2_model_names}}{

A character vector with 27 elements.
}
}
\usage{
loca2_model_names
}
\description{
LOCA2 model names scraped from the
\href{https://cirrus.ucsd.edu/~pierce/LOCA2/NAmer/}{LOCA2 server}.
}
\keyword{datasets}
