% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_temp_projection.R
\name{mean_temp_projection}
\alias{mean_temp_projection}
\title{Calculate yearly mean temperatures from temperature rasters}
\usage{
mean_temp_projection(tmax_rast, tmin_rast)
}
\arguments{
\item{tmax_rast}{terra raster stack of tmax values that must have terra::time
values containing date information to subset by year}

\item{tmin_rast}{terra raster stack of tmin values that must have terra::time
values containing date information to subset by year}
}
\value{
raster stack of yearly mean temperatures
}
\description{
Calculate yearly mean temperatures from temperature rasters
}
