% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quebec_farms}
\alias{quebec_farms}
\title{Quebec, Canada Maple Farm Addresses}
\format{
\subsection{\code{quebec_farms}}{

A dataframe with 85 rows and 4 columns:
\describe{
\item{farm}{Farm name}
\item{address}{Farm address}
\item{state}{Province}
\item{region}{Province region}
}
}
}
\usage{
quebec_farms
}
\description{
Maple syrup farm addresses scraped from
\href{https://www.bonjourquebec.com/en-us/to-see-and-do/delicious-discoveries/sugar-shacks}{bonjourquebec.com}.
}
\keyword{datasets}
