% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_day.R
\name{sap_day}
\alias{sap_day}
\title{Calculate ideal sap days from temperature rasters}
\usage{
sap_day(tmax_rast, tmin_rast, t_upper = 2.2, t_lower = -1.1, years = NULL)
}
\arguments{
\item{tmax_rast}{terra raster stack of tmax values that must have terra::time
values containing date information to subset by year}

\item{tmin_rast}{terra raster stack of tmin values that must have terra::time
values containing date information to subset by year}

\item{t_upper}{upper temperature value for the freeze/thaw cycle that
will be used in a logical statement to find ideal sap tapping days}

\item{t_lower}{lower temperature value for the freeze/thaw cycle that
will be used in a logical statement to find ideal sap tapping days}

\item{years}{integer vector of years to subset the raster stacks by. All
of the years must be present in the raster stacks}
}
\value{
list of length two - a raster stack of the proportion of ideal
sap tapping days per year, and a raster stack of the sum of ideal sap
tapping days per year
}
\description{
Maple sap flows the most on days where temperatures are oscillating between
freezing and thawing, known as the freeze/thaw cycle. This function uses
tmax and tmin rasters (potentially loaded in from \code{\link[=loca_t_rast]{loca_t_rast()}} or
\code{\link[=prism_t_rast]{prism_t_rast()}}) to calculate the proportion and sum of days in each year
that are ideal for maple sap tapping. Using a 30-year record of sap flow
and weather observations at the Uihlein Forest, Lake Placid, NY, sap flow
occurred on 80\% of the days when the minimum temperature fell below
-1.1 °C and maximum temperature exceeded 2.2 °C, which is where the
default values for the temperature threshold come from (Skinner et al.
2010).
}
\examples{
\dontrun{
# Load in rasters
# Assuming you used the download_loca2() function and placed the output
# in a folder called LOCA2
loca_rast <- loca_t_rast("LOCA2/ACCESS-CM2/0p0625deg/r1i1p1f1/")

# Calculate yearly sap days
k_upper <- 2.2 + 273.15
k_lower <- -1.1 + 273.15
loca_sap_day <- sap_day(loca_rast$tmax, loca_rast$tmin,
                        t_upper = k_upper, t_lower = k_lower)
}
}
