% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_farms.R
\name{sens_farms}
\alias{sens_farms}
\title{Get an sf dataframe with significance proportions by shapefile}
\usage{
sens_farms(
  farms_coords,
  sap_prop,
  elevation = NULL,
  shapefile,
  group_var,
  sig_var = "sens"
)
}
\arguments{
\item{farms_coords}{sf dataframe containing farms and their coordinates}

\item{sap_prop}{terra SpatRaster containing yearly sap day proportion
rasters, probably calculated from \code{\link[=sap_day]{sap_day()}}}

\item{elevation}{one layer terra SpatRaster containing elevation values}

\item{shapefile}{sf dataframe containing geographical shape boundaries}

\item{group_var}{character vector of length 1 containing the name of the
variable within the shapefile dataframe to group by}

\item{sig_var}{character vector of length 1 containing the name of the
significance column that will be averaged per shapefile region. This will
either be "sens" or "mk"}
}
\value{
sf dataframe that contains the significance averages for each group
and the corresponding shapefile boundaries - this can be plotted with
something like ggplot
}
\description{
This is a wrapper function that combines both \code{\link[=get_sens_farms]{get_sens_farms()}} and
\code{\link[=get_sens_joined]{get_sens_joined()}}. This function takes an sf dataframe of (assumed) farm
coordinates. Then, it extracts the time series of yearly sap day
proportions for each farm. Sen's slope is calculated for each time series.
The slope estimate and p-value are added as columns to the extracted
dataframe. Then, using the previously calculated
Sen's slope already present in the dataframe, a significance variable is
created that is 1 if the slope is significant-positive, -1 if the slope is
negative-significant, and 0 if the slope is not significant.
The farms are grouped by variable in the shapefile, and the average of the
significance value is taken for each group. A dataframe is returned that
contains the significance averages for each group and the corresponding
shapefile boundaries - this can be plotted with something like ggplot.
}
