% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_slope.R
\name{sens_slope}
\alias{sens_slope}
\title{Calculate the Sen's slope for a numeric vector}
\usage{
sens_slope(x, conf_level = 0.95)
}
\arguments{
\item{x}{numeric vector}

\item{conf_level}{numeric confidence level}
}
\value{
list of five items: Sen's slope estimate,
z statistic, p-value, sample size, and confidence interval
}
\description{
Calculate the Sen's slope for a numeric vector
}
\examples{
x <- rnorm(1000)
sens_slope(x)
}
