% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_slope_rast.R
\name{sens_slope_rast}
\alias{sens_slope_rast}
\title{Calculate the Sen's slope for every pixel of a raster}
\usage{
sens_slope_rast(t_rast, cores = 1)
}
\arguments{
\item{t_rast}{terra raster stack}

\item{cores}{integer amount of cores for parallelization}
}
\value{
terra raster stack with six layers: Sen's slope estimate,
z statistic, p-value, sample size, confidence interval lower bound, and
confidence interval upper bound
}
\description{
Calculate the Sen's slope for every pixel of a raster
}
\examples{
test_loca_file <- system.file("extdata", "test_loca_sap_day.tif",
                              package = "mapler")
sap_prop <- terra::rast(test_loca_file)
sens_slope_rast(sap_prop)
}
