% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_farms}
\alias{us_farms}
\title{U.S. Maple Farm Addresses}
\format{
\subsection{\code{us_farms}}{

A dataframe with 374 rows and 4 columns:
\describe{
\item{farm}{Farm name}
\item{address}{Farm address}
\item{state}{U.S. state}
\item{region}{U.S. state region}
}
}
}
\usage{
us_farms
}
\description{
Maple syrup farm addresses scraped from
\href{https://www.maplesyrupfarms.org/}{maplesyrupfarms.org}. Since scraping
the farm names was complicated, a few farm names are cut off at the first
hyphen.
}
\keyword{datasets}
