% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{U.S. State Abbreviations}
\format{
\subsection{\code{us_states}}{

A character vector with 50 elements.
}
}
\usage{
us_states
}
\description{
U.S. State Abbreviations
}
\keyword{datasets}
