% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnd.R
\name{download_ghcnd}
\alias{download_ghcnd}
\title{Download GHCNd files with given ids}
\usage{
download_ghcnd(ids, out_dir = paste0(getwd()))
}
\arguments{
\item{ids}{a character vector of GHCNd station ids}

\item{out_dir}{name of the folder where the files will be downloaded to}
}
\description{
Download GHCNd files with given ids
}
