% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sens_joined.R
\name{get_sens_joined}
\alias{get_sens_joined}
\title{Get an sf data frame of maple farms' Sen's slope grouped by shapefile
variable}
\usage{
get_sens_joined(farms_sf, shapefile, variable)
}
\arguments{
\item{farms_sf}{sf dataframe containing the time series of sap day
proportions for each year, the geometry, the Sen's slope estimate,
the Sen's slope p-value, and the significance value,
probably outputted by \code{\link[=get_sens_significance]{get_sens_significance()}}}

\item{shapefile}{sf dataframe containing geographical shape boundaries}

\item{variable}{character vector of length 1 containing the name of the
variable within the shapefile dataframe to group by}
}
\value{
sf dataframe that contains the significance averages for each group
and the corresponding shapefile boundaries - this can be plotted with
something like ggplot
}
\description{
This function takes an sf dataframe probably outputted by \code{\link[=get_sens_farms]{get_sens_farms()}}
or \code{\link[=get_sens_significance]{get_sens_significance()}}.
The farms are grouped by variable in the shapefile, and the average of the
significance value is taken for each group. A dataframe is returned that
contains the significance averages for each group and the corresponding
shapefile boundaries - this can be plotted with something like ggplot.
}
