% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sens_joined.R
\name{get_sens_joined}
\alias{get_sens_joined}
\title{Get an sf data frame of maple farms' Sen's slope grouped by shapefile
variable}
\usage{
get_sens_joined(farms_sf, shapefile, group_var, sig_var = "sens_significant")
}
\arguments{
\item{farms_sf}{sf dataframe containing the time series of sap day
proportions for each year, the geometry, the Sen's slope estimate,
the Sen's slope p-value, and the significance value,
probably outputted by \code{\link[=get_sens_farms]{get_sens_farms()}}}

\item{shapefile}{sf dataframe containing geographical shape boundaries}

\item{group_var}{character vector of length 1 containing the name of the
variable within the shapefile dataframe to group by}

\item{sig_var}{character vector of length 1 containing the name of the
significance column in farms_sf that will be averaged per shapefile region}
}
\value{
sf dataframe that contains the significance averages for each group
and the corresponding shapefile boundaries - this can be plotted with
something like ggplot
}
\description{
This function takes an sf dataframe probably outputted by \code{\link[=get_sens_farms]{get_sens_farms()}}.
The farms are grouped by variable in the shapefile, and the average of the
significance value is taken for each group. A dataframe is returned that
contains the significance averages for each group and the corresponding
shapefile boundaries - this can be plotted with something like ggplot.
}
\examples{
\dontrun{
# Read in farm coordinates and sap day projection
farms_sf <- sf::st_as_sf(farms_coords, coords = c("long", "lat"), crs = 4326)
test_loca_file <- system.file("extdata", "test_loca_sap_day.tif",
                              package = "mapler")
sap_prop <- terra::rast(test_loca_file)

# Get Sen's slope for every farm
farms_sf <- get_sens_farms(farms_sf, sap_prop)

# Read in eco regions shape file
shapefile <- sf::read_sf("Data_Clean/NA_Eco_Level3/NA_CEC_Eco_Level3.shp")
variable <- names(shapefile)[2]

# Get mean significance by eco region
get_sens_joined(farms_sf, shapefile, variable)
}
}
